Create Mock folder and move the mcks files
